refactor(cart): move store into feature slice
