+++ src/features/cart/model/index.js
@@ -1,0 +1,2 @@
+export { default as useCartStore } from './cartStore';
+export { selectCartCount, selectCartItems, selectCartTotal } from './selectors';
