+++ src/widgets/layout/header/ui/Header.jsx
@@ -6,10 +6,8 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useMemo, useState } from 'react';
 
-import { selectCartCount } from '@/app/store/cartSelectors';
-import useCartStore from '@/app/store/cartStore';
-import { cx } from '@/lib/utils/cx';
-import { useSmartHeader } from '@/lib/utils/useSmartHeader';
+import { selectCartCount, useCartStore } from '@/features/cart/model';
+import { cx, useSmartHeader } from '@/shared/lib';
 
 const NAV = [
   { href: '/products', label: 'Каталог' },
